+++ src/main/java/com/umar/simply/jdbc/meta/Table.java
@@ -3,14 +3,14 @@ package com.umar.simply.jdbc.meta;
 public class Table {
 
     private final String tableName;
-    private final Column<Long> idColumn;
+    private final Column<Integer> idColumn;
 
-    private Table(String tableName, Column<Long> idColumn) {
+    private Table(String tableName, Column<Integer> idColumn) {
         this.tableName = tableName;
         this.idColumn = idColumn;
     }
 
-    public static Table table(String tableName, Column<Long> idColumn){
+    public static Table table(String tableName, Column<Integer> idColumn){
         return new Table(tableName, idColumn);
     }
 
@@ -18,11 +18,11 @@ public class Table {
         return tableName;
     }
 
-    public Column<Long> getIdColumn() {
+    public Column<Integer> getIdColumn() {
         return idColumn;
     }
 
-    public static Table as(String tableName, String alias, Column<Long> idColumn) {
+    public static Table as(String tableName, String alias, Column<Integer> idColumn) {
         return new Table(tableName +" "+alias, idColumn);
     }
